fixing so that the new top header is inside with all the draggable elements
