+++ TrafikkApp/components/helpers/useMeasure.js
@@ -1,0 +1,17 @@
+// import { useState, useCallback } from 'react';
+
+// export const useMeasure = (deps) => {
+//     const [rect, setRect] = useState(null);
+
+//     const ref = useCallback(
+//         (node) => {
+//             console.log(node);
+//             // if (node !== null) {
+//             //     setRect(node.getBoundingClientRect());
+//             // }
+//         },
+//         [deps]
+//     );
+
+//     return [rect, ref];
+// };
